requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
apscheduler=3.6.3=pypi_0
ca-certificates=2020.12.8=hecd8cb5_0
certifi=2020.12.5=py37hecd8cb5_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
cryptography=3.3.1=pypi_0
decorator=4.4.2=pypi_0
idna=2.10=pypi_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
ncurses=6.2=h0a44026_1
openssl=1.1.1i=h9ed2024_0
pip=20.3.3=py37hecd8cb5_0
pycparser=2.20=pypi_0
pytelegrambotapi=3.7.4=pypi_0
python=3.7.9=h26836e1_0
python-telegram-bot=13.1=pypi_0
pytz=2020.5=pypi_0
readline=8.0=h1de35cc_0
requests=2.25.1=pypi_0
setuptools=51.0.0=py37hecd8cb5_2
six=1.15.0=pypi_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.26.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
